Add zooming capabilities
This should address the issue in https://github.com/BigFatDog/cubism-es/issues/9.

There is two part in the change itself one to allow the user to interact
with the graph to indicate wich area of the graph should be zoomed and
one to actually to allow redrawing the graph with zoomed data.

The data model for cubism didn't map very well with the later, metrics
are meant to be appended on and on and not rewind back in time a replot
new (more detailed metrics).

So this required a bit more change.

In order to use it add this in the code:
```
    const z = d3.select("body").append("div")
        .attr("class", "zoom");
    context.zoom(function(start, end) {
        console.log(`Doing a zoom from point ${start} to point ${end}`);
        context.zoom().zoomTime(start, end);
    }).render(z);
```