Improve config for rollup
Get rid of a warning related to babel about using an implicit value.
Remove warning about circular deps in the dev config, it was already
done in the usual one.

Change the folders that trigger the reload in the dev config, when
I first created the file I took inspiration of what was done for
a grafana plugin but here it's better to monitor 2 folders.