Do not display value(s) that are null
In theory there is provision not to display values that are not a number
because there is a test `isNaN(value)? "": _format` but it seems that
null is a number and so `NaN` is displayed.

I changed `simple.html` to have the last value be null and also removed
a div that seems not to be needed