Allow other to control how they import `cubism-es`
We build multiple things:
* a CJS "module"
* a ES module
* a JS standalone file (mostly for testing but it might be useful for
  someone else).

But there was no control on how it was imported when using NodeJS, which
mean that it's almost certain that it's the CJS file that will be used.

This seems to work most of the time but somehow when `cubism-es` is used
elsewhere and you want to test it using jest there is an issue with d3
global variable. If I try to `require` `d3` in `jest-setup.js` then

```
import * as cubism from `cubism-es`
```

Is not quite working because cubism is just an empty object.
So instead with the new export we can import the `ESM` version of
`cubism-es` and avoid the issue with the global variable.