Improve the way rendering of value is done
Instead of using a complicated logic that iterate between
size_of(colors)/2 and size_of(colors) time on the data array do it in
one pass. Also avoid complicated range change with negative values,
instead use a simplier model where we map the value to one between
0 and horizon_height*size_of(colors) / 2 and use modulo + and integer to
find the right color and the height of the bar within this color.

Signed-off-by: Matthieu Patou <mat@matws.net>